+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. This change
 - Implement `console.time*` functionality ([#997](https://github.com/planck-repl/planck/issues/997))
 
 ### Changed
+- Update to ClojureScript 1.10.597
 - Ship with `com.cognitect/transit-js` supporting meta serialization functionality ([#1009](https://github.com/planck-repl/planck/issues/1009))
 - Remove unneeded `planck.repl` require in `planck.io` and `planck.http` ([#1007](https://github.com/planck-repl/planck/issues/1007))